% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate_exprset}
\alias{annotate_exprset}
\title{Annotate Probesets of ExpressionSet}
\usage{
annotate_exprset(.exprs)
}
\arguments{
\item{.exprs}{A \link[Biobase:class.ExpressionSet]{Biobase::ExpressionSet} to annotate}
}
\value{
A \link[Biobase:class.ExpressionSet]{Biobase::ExpressionSet} with \link[Biobase:featureData]{Biobase::featureData} added incorporating existing chip annotation.
}
\description{
Add \link[Biobase:featureData]{Biobase::featureData} to a \link[Biobase:class.ExpressionSet]{Biobase::ExpressionSet} consisting of probeset annotation.
}
\details{
This function is inspired by affycoretools. If the standard approach when building an
ExpressionSet (from affy data) is to include annotation then downstream work is much easier. This
is that approach.
}
\note{
There is a great discussion in affycoretools about the problem of one-to-many and many-to-many
mapping of annotation. This package uses the first entry approach, which is not ideal but is practical.
}
\examples{
\dontrun{
annotate_exprset(exprs)
}
}
