% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate_with}
\alias{annotate_with}
\alias{annotate_with_symbol}
\alias{annotate_with_ensembl}
\alias{annotate_with_entrezid}
\title{Generate Map of Probeset Annotation}
\usage{
annotate_with(.exprs, with = NULL, use = c("first", "last", "everything"))

annotate_with_symbol(...)

annotate_with_ensembl(...)

annotate_with_entrezid(...)
}
\arguments{
\item{.exprs}{An \link[Biobase:class.ExpressionSet]{Biobase::ExpressionSet} or \link[AnnotationDbi:AnnotationDb-class]{AnnotationDbi::ChipDb} object}

\item{with}{The fields to map from probesets}

\item{use}{The selection process when more than one element matches. See \code{\link[=selectors]{selectors()}}
for details.}

\item{...}{Any further arguments to the function}
}
\value{
A named vector or data frame with names/rownames as probesets
and the mapped data as values.
}
\description{
Generate Map of Probeset Annotation
}
\details{
This is a composite function. In its simplest form, it
will provide a map of probesets to an annotation fields (see
\code{\link[=annotation_fields]{annotation_fields()}} for fields). It does a simple approach defaulted
to in \link[AnnotationDbi:AnnotationDb-class]{AnnotationDbi::mapIds} and selects the first annotation. This map
will be a named vector with the name being the probeset and the value
being the mapped field.

If multiple fields are provided, then this process is repeated for all
fields and the result is a data frame with rownames as probeset and each
field being a separate column. Note that the "take-the-first" approach
still holds for each column, which could provide some surprising results.
}
\section{Functions}{
\itemize{
\item \code{annotate_with_symbol}: Annotate with gene symbol

\item \code{annotate_with_ensembl}: Annotate with Ensembl ID

\item \code{annotate_with_entrezid}: Annotate with entrez gene ID
}}

