% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotation_fields}
\alias{annotation_fields}
\title{Return Available Annotation Fields}
\usage{
annotation_fields(.exprs)
}
\arguments{
\item{.exprs}{A \link[Biobase:class.ExpressionSet]{Biobase::ExpressionSet} or \link[AnnotationDbi:AnnotationDb-class]{AnnotationDbi::ChipDb}}
}
\value{
The list of columns in the annotation data.
}
\description{
Return Available Annotation Fields
}
