% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{.map_fields_121}
\alias{.map_fields_121}
\title{Generate Map of Annotation}
\usage{
.map_fields_121(a, fields = NULL, selector = .select_first_element)
}
\arguments{
\item{a}{An \link[AnnotationDbi:AnnotationDb-class]{AnnotationDbi::ChipDb} object.}

\item{fields}{The fields to extract}

\item{selector}{A selector when multiple matches are found}
}
\value{
A list or data frame of mapped data with names being the probesets.
}
\description{
Maps annotation data from probeset to other columns but
chooses only one match.
}
\details{
This is an annotation function that maps from probesets to
other fields. It selects only one mapping, even if multiple mappings
exist. The selector parameter determines how that map is chosen.
}
