% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{.map_fields_12many}
\alias{.map_fields_12many}
\title{Generate One-to-Many Map of Probeset Annotation}
\usage{
.map_fields_12many(.exprs, fields = NULL)
}
\arguments{
\item{.exprs}{An ExpressionSet or Annotation Object}

\item{fields}{The fields to extract}
}
\value{
A data frame of probes by fields with many values embedded as lists.
}
\description{
Provides a mapping of annotation as embedded lists.
}
\details{
This is a somewhat useful function, but be careful what you wish for. This
function provides the one-to-many mapping of annotation via a data frame
full of lists. That is, for a given probe and a given field it will be
a list of values. For something like accession numbers this could be a list
of 50 accessions. Embedded within a single position in the data frame of
probeset x field. Still, if you want that level of detail then this is the
object for you.
}
