% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apt.R
\name{extract_apt_platform_string}
\alias{extract_apt_platform_string}
\title{Extract platform info from APT header}
\usage{
extract_apt_platform_string(s)
}
\arguments{
\item{s}{A vector of strings representing APT parameters}
}
\value{
A string representing the extracted string type
}
\description{
Given a vector of strings, extract the platform or chip type.
}
\details{
In the APT output, the parameter

\if{html}{\out{<div class="sourceCode">}}\preformatted{affymetrix-algorithm-param-apt-opt-chip-type
}\if{html}{\out{</div>}}

describes the Affymetrix platform (chip) used in the analysis. This routine will
find this entry in the vector and extract it.
}
\examples{
\dontrun{
extract_apt_platform_string(c("#\%affymetrix-algorithm-param-apt-opt-chip-type=Clariom-D"))
}
}
