% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_by_variance}
\alias{filter_by_variance}
\title{Filter out genes (features) with low variability}
\usage{
filter_by_variance(.exprs, threshold = 0)
}
\arguments{
\item{.exprs}{the ExpressionSet to use}

\item{threshold}{a minimum acceptable threshold of gene expression variance}
}
\value{
A new ExpressionSet reduced to genes passing the criteria.
}
\description{
Filter out genes (features) with low variability
}
\examples{
\dontrun{
filter_by_variance(exprset, threshold = 0.5)
}
}
