% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_low_expression}
\alias{filter_low_expression}
\title{Filter out genes (features) with no samples above noise (low expression)}
\usage{
filter_low_expression(.exprs, threshold = 5)
}
\arguments{
\item{.exprs}{the ExpressionSet to use}

\item{threshold}{a minimum acceptable threshold of gene expression (default 5)}
}
\value{
A new ExpressionSet reduced to genes passing the criteria.
}
\description{
Filter out genes (features) with no samples above noise (low expression)
}
