% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_low_range}
\alias{filter_low_range}
\title{Filter genes with range less than threshold}
\usage{
filter_low_range(.exprs, threshold = 1)
}
\arguments{
\item{.exprs}{ExpressionSet to use for filtering.}

\item{threshold}{Range to use as minimum (default 1)}
}
\value{
A new ExpressionSet with low range genes removed.
}
\description{
Filter genes with range less than threshold
}
