% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_n_above_threshold}
\alias{filter_n_above_threshold}
\title{Filter genes on number of samples above a target threshold}
\usage{
filter_n_above_threshold(.exprs, threshold = 5, n = 0)
}
\arguments{
\item{.exprs}{ExpressionSet}

\item{threshold}{A threshold to consider low}

\item{n}{The number of samples exceeding threshold}
}
\value{
ExpressionSet reduced to rows with more than \code{n} columns larger than \code{threshold}.
}
\description{
Filter genes on number of samples above a target threshold
}
\note{
If n is fractional, it is treated as a percentage.
}
\examples{
\dontrun{
# Filter probesets with more than 20\% above expression level of 5.
filter_n_above_threshold(exprset, threshold = 5, n = 0.2)

# Filter probesets with more than 10 samples above expression level of 10.
filter_n_above_threshold(exprset, threshold = 10, n = 10)
}
}
