% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apt.R
\name{index_of_last_apt_comment}
\alias{index_of_last_apt_comment}
\title{Find row number of last comment in vector of strings}
\usage{
index_of_last_apt_comment(s)
}
\arguments{
\item{s}{A vector of strings}
}
\value{
The row number of the last comment
}
\description{
Find row number of last comment in vector of strings
}
