% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.R
\name{is.ExpressionSet}
\alias{is.ExpressionSet}
\title{Is object an ExpressionSet}
\usage{
is.ExpressionSet(.exprs)
}
\arguments{
\item{.exprs}{The object}
}
\value{
Logical if .exprs is an ExpressionSet
}
\description{
Is object an ExpressionSet
}
\examples{
\dontrun{
is.ExpressionSet(affy::rma(Dilution))
}
}
