% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apt.R
\name{load_apt_rma}
\alias{load_apt_rma}
\title{Load APT data into ExpressionSet}
\usage{
load_apt_rma(basename = "rma")
}
\arguments{
\item{basename}{The file stem to use for loading the summary and report txt files}
}
\value{
A [\code{\link[Biobase:class.ExpressionSet]{Biobase::ExpressionSet()}}] object representing the contents of the file(s).
}
\description{
Parse and load APT output into a [\code{\link[Biobase:class.ExpressionSet]{Biobase::ExpressionSet()}}] object.
}
\details{
APT produces two output files of interest: the report and summary files. The
report file consists of a series of lines of parameter settings (prefixed with \verb{#\%})
and rows of statistics about the CEL file that was processed. The summary file starts
with the parameter settings (same as report file) and includes the gene expression estimates.

This function will load both files (using \code{basename} as the file stem for both the
summary and report files). The information will be stored in a [\code{\link[Biobase:class.ExpressionSet]{Biobase::ExpressionSet()}}]
object. There are a couple of points about the object:
\itemize{
\item \code{assayData} will contain the gene expression values
\item \code{annotation} will contain the chip name from the metadata
\item \code{pData} will contain the statistics about the chip (average intensity, etc)
\item \code{preproc(experimentData())} contains the algorithm parameters used in the APT process
}
}
\examples{

\dontrun{
# Load rma.summary.txt and rma.report.txt
load_apt_rma(basename = "rma")
}
}
