% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apt.R
\name{load_apt_rma_report}
\alias{load_apt_rma_report}
\title{Parse and return the report file from APT}
\usage{
load_apt_rma_report(f)
}
\arguments{
\item{f}{A file with the report content.}
}
\value{
A list with algorithm parameters (\code{parameters}) and normalization statistics (\code{statistics}).
}
\description{
Parse and return the report file from APT
}
