% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apt.R
\name{load_apt_rma_summary}
\alias{load_apt_rma_summary}
\title{Load the APT RMA summary file}
\usage{
load_apt_rma_summary(f)
}
\arguments{
\item{f}{The file containing RMA summary information}
}
\value{
A [\code{\link[Biobase:class.ExpressionSet]{Biobase::ExpressionSet()}}] with the expression data (assayData) and experimentInfo.
}
\description{
Load the APT RMA summary file
}
