% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apt.R
\name{parse_apt_header}
\alias{parse_apt_header}
\title{Extract header lines from a vector of strings}
\usage{
parse_apt_header(hdr)
}
\arguments{
\item{hdr}{A vector of strings to examine}
}
\value{
A list of header name=value pairs
}
\description{
Extract header lines from a vector of lines from a file.
}
\details{
APT headers consist of the prefix \verb{#\%}. This routine finds these
lines, strips the prefix and returns the result as a list.
}
