% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{selectors}
\alias{selectors}
\alias{.select_first_element}
\alias{.select_last_element}
\alias{.select_longest_element}
\title{Select an annotation element}
\usage{
.select_first_element(.x)

.select_last_element(.x)

.select_longest_element(.x)
}
\arguments{
\item{.x}{A list of character vectors to select from.}
}
\description{
Given a list of annotation elements, select the most
representative one.
}
\details{
In one-to-many mappings, there are many ways to pick a representative
element. It could be the first, last, etc. These routines are selectors that
implement these various logical choices.
}
\section{Functions}{
\itemize{
\item \code{.select_first_element}: Select the first element.

\item \code{.select_last_element}: Select the last element.

\item \code{.select_longest_element}: Select the element with the longest string length.
}}

