% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset.R
\name{subset_by}
\alias{subset_by}
\alias{subset_by_symbol}
\title{Subset an ExpressionSet by a list of identifiers}
\usage{
subset_by(.exprs, subset = NULL, by = c("PROBESET", "SYMBOL"), verbose = TRUE)

subset_by_symbol(...)
}
\arguments{
\item{.exprs}{An ExpressionSet to subset}

\item{subset}{A list of probeset identifiers, or other identifier (e.g., gene symbol)}

\item{by}{A selector of the mapping (default is PROBESET, choices include SYMBOL).}

\item{verbose}{Should statistics be reported on what is subset.}

\item{...}{Any additional arguments}
}
\value{
A new ExpressionSet with fewer rows, representing a subset.
}
\description{
Subset an ExpressionSet by providing a list of probe/gene identifiers.
}
\details{
This function subsets an ExpressionSet by probe (row). This can be done
directly from the \link[Biobase:class.ExpressionSet]{Biobase::ExpressionSet} object, however there are a few additional
features of this function.

First, you can use other fields (like SYMBOL) which are present in annotation to do
a subsetting. It handles the mappings back and forth to probesets.

Second, it tells you what it's doing. Not all of your identifiers may be present (this
happens in gene signatures, etc). So it will tell you that it's happening (with \code{verbose=TRUE}).
}
\section{Functions}{
\itemize{
\item \code{subset_by_symbol}: Subset by gene symbol
}}

\examples{
\dontrun{
# Subset by probesets
subset_by(exprs, subset=c("1007_s_at","1008_s_at"))
# Subset by gene
subset_by(exprs, subset=c("STAT1","EGFR","TP53"), by = "SYMBOL")
}
}
