% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize}
\alias{summarize}
\alias{.summarize_maxmedian}
\alias{.summarize_minmedian}
\alias{.summarize_median}
\alias{.summarize_mean}
\alias{.summarize_tukey_biweight}
\title{Summarize a group of probesets}
\usage{
.summarize_maxmedian(i, x)

.summarize_minmedian(i, x)

.summarize_median(i, x)

.summarize_mean(i, x)

.summarize_tukey_biweight(i, x)
}
\arguments{
\item{i}{The rows of \code{x} to use in summarizing}

\item{x}{The matrix of expression (rows=genes, columns=samples)}
}
\value{
A summarized matrix with same number of columns, but one row.
}
\description{
Given a set of samples and a set of probesets, summarize to one measurement
per sample.
}
\details{
Summarizing data from multiple probesets to a single value can be tricky since not
all probesets are equally good.
}
\section{Functions}{
\itemize{
\item \code{.summarize_maxmedian}: Use the probeset with the max median expression.

\item \code{.summarize_minmedian}: Use the probeset with the min median expression.

\item \code{.summarize_median}: Use the median value

\item \code{.summarize_mean}: Use the mean value

\item \code{.summarize_tukey_biweight}: Use Tukey's BiWeight calculation
}}

