% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize_by}
\alias{summarize_by}
\alias{summarize_by_symbol}
\alias{summarize_by_entrezid}
\alias{summarize_by_ensembl}
\title{Summarize ExpressionSet rows to reduced form}
\usage{
summarize_by(
  .exprs,
  by = "SYMBOL",
  method = c("median", "maxmedian", "mean", "minmedian", "tukey.biweight")
)

summarize_by_symbol(...)

summarize_by_entrezid(...)

summarize_by_ensembl(...)
}
\arguments{
\item{.exprs}{An ExpressionSet to summarize}

\item{by}{The annotation field to summarize by (e.g., SYMBOL)}

\item{method}{The method to use for summarization. Can include median, max, mean, min, tukey.biweight.}

\item{...}{Any additional parameters}
}
\value{
A summarized ExpressionSet containing phenodata, etc from the original ExpressionSet
but with summarize rows.
}
\description{
Summarize an ExpressionSet to a reduced form by grouping
on a feature annotation variable.
}
\details{
We often need to summarize an ExpressionSet by Gene Symbol or other
annotation field. There are two parts to having to do this work. First, you
have to map/group the probesets by this annotation field. Then you have to
pick a summarization method. This function does those two things and
returns an ExpressionSet with the reduced information, copying where
possible the attributes of the original ExpressionSet.
}
\section{Functions}{
\itemize{
\item \code{summarize_by_symbol}: Summarize by gene symbol.

\item \code{summarize_by_entrezid}: Summarize by entrez gene ID.

\item \code{summarize_by_ensembl}: Summarize by Ensemble ID.
}}

\examples{
\dontrun{
summarize_by(exprset, by = "SYMBOL", method = "max")
}
}
