% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{translate_ExpressionSet}
\alias{translate_ExpressionSet}
\title{Translate ExpressionSet to alternate dimensions.}
\usage{
translate_ExpressionSet(
  src,
  new_exprs,
  pdata = NULL,
  fdata = NULL,
  annotation = NULL
)
}
\arguments{
\item{src}{Original ExpressionSet to clone}

\item{new_exprs}{A new matrix of expression data (with either fewer rows or columns)}

\item{pdata}{Potentially new phenoData to use}

\item{fdata}{Potentially new featureData to use}

\item{annotation}{Potentiall new annotation}
}
\value{
A new expression set
}
\description{
An ExpressionSet can be morphed into a new dimensionality, but only carefully.
}
\details{
An ExpressionSet needs to maintain the correct dimensions for the linked data
within it. This function attempts to clone an ExpressionSet then alter one of the dimensions
based on input data. All other data is taken from the original copy.
}
\examples{
\dontrun{
translate_ExpressionSet(exprset, matrix(c(1,2,3)))
}
}
