% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{summarize}
\alias{summarize}
\alias{.summarize_max}
\alias{.summarize_min}
\alias{.summarize_median}
\alias{.summarize_mean}
\title{Summarize a group of probesets}
\usage{
.summarize_max(i, x)

.summarize_min(i, x)

.summarize_median(i, x)

.summarize_mean(i, x)
}
\arguments{
\item{i}{The rows of \code{x} to use in summarizing}

\item{x}{The matrix of expression (rows=genes, columns=samples)}
}
\value{
A summarized matrix with same number of columns, but one row.
}
\description{
Given a set of samples and a set of probesets, summarize to one measurement
per sample.
}
\details{
Summarizing data from multiple probesets to a single value can be tricky since not
all probesets are equally good.
}
\section{Functions}{
\itemize{
\item \code{.summarize_max}: Use the max value

\item \code{.summarize_min}: Use the min value

\item \code{.summarize_median}: Use the median value

\item \code{.summarize_mean}: Use the mean value
}}

